backend/requirements.txt
fastapi
uvicorn[standard]
httpx
scikit-learn
spacy
vaderSentiment
pydantic
python-dotenv
beautifulsoup4
pytest
pytest-asyncio
